requirements.txt
selenium>=4.0.0
requests>=2.28.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
urllib3>=1.26.0